Add super basic entry expiry.
It expires keys on fetch. In a production system it would be nice when adding
new entries to check if there's any expired to make space for the new one as
opposed to automatically evict the oldest one (which could still be unexpired).

That would require an extra data structure like a priority queue and would
change the runtime for cache operations.